Add Brew Function
