+++ .idea/misc.xml
@@ -4,6 +4,7 @@
     <option name="filePathToZoomLevelMap">
       <map>
         <entry key="..\:/Users/giaco/OneDrive/Desktop/progetto-birra-1-beer2beer/app/src/main/res/layout/fragment_add_recipe.xml" value="0.24615384615384617" />
+        <entry key="..\:/Users/giaco/OneDrive/Desktop/progetto-birra-1-beer2beer/app/src/main/res/layout/fragment_brew.xml" value="0.24615384615384617" />
         <entry key="..\:/Users/giaco/OneDrive/Desktop/progetto-birra-1-beer2beer/app/src/main/res/layout/fragment_get_started.xml" value="0.25" />
         <entry key="..\:/Users/giaco/OneDrive/Desktop/progetto-birra-1-beer2beer/app/src/main/res/layout/fragment_recipe_detail.xml" value="0.25" />
         <entry key="..\:/Users/giaco/OneDrive/Desktop/progetto-birra-1-beer2beer/app/src/main/res/layout/recipe_ingredient_item.xml" value="0.25" />
