+++ app/src/main/java/com/example/beer2beer/SharedViewModel.kt
@@ -1,12 +1,16 @@
 package com.example.beer2beer
 
 import android.app.Application
+import android.util.Log
 import androidx.lifecycle.AndroidViewModel
+import androidx.lifecycle.LifecycleOwner
+import androidx.lifecycle.LiveData
 import androidx.lifecycle.viewModelScope
 import com.example.beer2beer.database.AppDatabase
 import com.example.beer2beer.database.entities.Equipment
 import com.example.beer2beer.database.entities.Recipe
 import com.example.beer2beer.database.entities.RecipeHasIngredient
+import com.example.beer2beer.database.entities.RecipeIngredients
 import com.example.beer2beer.repository.EquipmentRepository
 import com.example.beer2beer.repository.IngredientRepository
 import com.example.beer2beer.repository.RecipeRepository
@@ -17,8 +21,6 @@ import org.apache.commons.math3.optim.linear.*
 import org.apache.commons.math3.optim.nonlinear.scalar.GoalType
 
 class SharedViewModel(application: Application) : AndroidViewModel(application) {
-
-    private val DEFAULT_MAX_ITER = MaxIter(100)
 
     // get the Database instance
     private val db = AppDatabase.getInstance(getApplication<Application>().applicationContext)
@@ -31,6 +33,7 @@ class SharedViewModel(application: Application) : AndroidViewModel(application) 
     val recipes = recipeRepository.getAllRecipes()
     val ingredients = ingredientRepository.getAllIngredients()
     val equipment = equipmentRepository.getAllEquipments()
+
 
 
     //RECIPES
@@ -77,67 +80,8 @@ class SharedViewModel(application: Application) : AndroidViewModel(application) 
         recipeRepository.deleteRecipe(id)
     }
 
-    fun whatShouldIBrewToday(): String {
-
-        var bestRecipe: Recipe? = null
-        var bestQuantity = 0.0
-        val ss = SimplexSolver()
-
-        viewModelScope.launch(Dispatchers.IO) {
-            recipes.value?.forEach { recipe ->
-                val recipeIngredients = db.recipeDao().getRecipeIngredients(recipe.id)
-                val constraint = ArrayList<LinearConstraint>()
-                val coeff = DoubleArray(recipeIngredients.size + 1)
-
-                for (i in recipeIngredients.indices) {
-                    coeff[i] = -recipeIngredients.get(i).ratio
-                    val c = DoubleArray(recipeIngredients.size + 1)
-                    c.fill(0.0)
-                    c[i] = 1.0
-                    if(recipeIngredients.get(i).name == "Water"){
-                        constraint.add(
-                            LinearConstraint(
-                                c,
-                                Relationship.LEQ,
-                                recipeIngredients.get(i).quantity * 1000
-                            )
-                        )
-                    }else{
-                        constraint.add(
-                            LinearConstraint(
-                                c,
-                                Relationship.LEQ,
-                                recipeIngredients.get(i).quantity
-                            )
-                        )
-                    }
-
-                }
-
-                coeff[coeff.lastIndex] = 1.0
-                constraint.add(LinearConstraint(coeff, Relationship.EQ, 0.0))
-
-                val c = DoubleArray(recipeIngredients.size + 1)
-                c.fill(0.0)
-                c[c.lastIndex] = 1.0
-
-                val fObb = LinearObjectiveFunction(c, 0.0)
-                val constr = LinearConstraintSet(constraint)
-
-                val solution =
-                    ss.optimize(
-                        DEFAULT_MAX_ITER, fObb, constr,
-                        GoalType.MAXIMIZE, NonNegativeConstraint(true)
-                    ).value
-
-                if (bestRecipe == null || solution > bestQuantity) {
-                    bestRecipe = recipe
-                    bestQuantity = solution
-                }
-            }
-        }
-
-        return bestRecipe?.name ?: ""
+    fun getRecipeIngredients(id: Int): LiveData<List<RecipeIngredients>> {
+        return recipeRepository.getRecipeIngredients(id)
     }
 
     fun updateIngredient(name: String, newQuantity: Double) {
