+++ app/src/main/java/com/example/beer2beer/database/entities/RecipeIngredients.kt
@@ -6,7 +6,6 @@ import androidx.room.PrimaryKey
 
 @Entity
 data class RecipeIngredients(
-
     val name: String,
     val ratio: Double,
     val quantity: Double
